last-change
-semua struktur data selain array sudah diganti
- terdapat beberapa perubahan untuk menyesuaikan instruksi soal seperti:
1. Jika input jumlah kosong (langsung enter) maka default jumlahnya 1
2. Hasil perhitungan berpajak (dengan koma) diganti dengan tipe data double

+++ KohiShop/Makanan.java
@@ -1,7 +1,7 @@
 // Alif
 public class Makanan extends Menu
 {
-    public Makanan(String kode, String nama, int harga) 
+    public Makanan(String kode, String nama, double harga) 
     {
         super(kode, nama, harga);
     }
@@ -9,6 +9,6 @@ public class Makanan extends Menu
     @Override
     public void tampilkanData()
     {
-        System.out.printf("| %-2s | %-33s | %-2d |\n", getKode(), getNama(), getHarga());
+        System.out.printf("| %-2s | %-33s | %-2.2f |\n", getKode(), getNama(), getHarga());
     }
 }
